Creating a Model or Database Table
You have your database connected now but need a table to hold certain data. You are creating an API for a bookstore. Each book has to be identified uniquely and possesses properties that help in proper categorization. Decide what the properties or fields of this Book Model will be and create a Database Table for it.